adds map to post show
