+++ app/views/posts/new.html.erb
@@ -35,7 +35,7 @@
     </div>
     <p class="d-none">Coordenadas</p>
     <div class="col-6">
-        <div id="map" style="height: 300px; width: 600px"></div>
+        <div id="map" style="height: 300px; width: 600px" data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>"></div>
     </div>
   </div>
 </div>
