+++ app/views/posts/show.html.erb
@@ -24,16 +24,25 @@
  </div>
   <!-- descripcion -->
   <div class=" container px-5">
-    <p><%= @post.description.capitalize %></p>
+    <div class="row">
+      <div class="col-6">
+        <h2>Descripción</h2>
+        <p><%= @post.description.capitalize %></p>
+        <br>
+        <h2>Caracteristicas</h2>
+        <p><%= @post.pet.animal.capitalize %></p>
+        <p><%= @post.pet.breed.capitalize %></p>
+        <p><%= @post.pet.size.capitalize %></p>
+        <p><%= @post.pet.gender.capitalize %></p>
+      </div>
+      <div class="col-6">
+        <div id="map"
+          style="width: 100%; height: 400px;"
+          data-marker="<%= @marker.to_json %>"
+          data-mapbox-api-key="<%= ENV['MAPBOX_API_KEY'] %>"></div>
+      </div>
+    </div>
   </div>
-  <h3 class="px-5">Caracteristicas</h3>
-  <div class="px-5">
-    <p><%= @post.pet.animal.capitalize %></p>
-    <p><%= @post.pet.breed.capitalize %></p>
-    <p><%= @post.pet.size.capitalize %></p>
-    <p><%= @post.pet.gender.capitalize %></p>
-  </div>
-
   <div>
     <%= render "contact_form", post: @post %>
   </div>
